+++ python-sdk/tests/files/type/test_json.py
@@ -4,7 +4,7 @@ from unittest import mock
 
 import pandas as pd
 
-from astro.dataframes.load_options import PandasLoadOptions
+from astro.dataframes.load_options import PandasJsonLoadOptions, PandasLoadOptions
 from astro.dataframes.pandas import PandasDataframe
 from astro.files.types import JSONFileType
 
@@ -26,6 +26,16 @@ def test_read_json_file_with_pandas_opts(mock_read_json):
     """Test pandas option get pass to read_json"""
     path = str(sample_file.absolute())
     json_type = JSONFileType(path, load_options=PandasLoadOptions(encoding="utf-8"))
+    with open(path) as file:
+        json_type.export_to_dataframe(file)
+    mock_read_json.assert_called_once_with(file, encoding="utf-8")
+
+
+@mock.patch("astro.files.types.json.pd.read_json")
+def test_read_json_file_with_pandas_opts_with_deprecated_load_options(mock_read_json):
+    """Test pandas option get pass to read_json"""
+    path = str(sample_file.absolute())
+    json_type = JSONFileType(path, load_options=PandasJsonLoadOptions(encoding="utf-8"))
     with open(path) as file:
         json_type.export_to_dataframe(file)
     mock_read_json.assert_called_once_with(file, encoding="utf-8")