+++ python-sdk/tests_integration/astro_deploy/deploy.sh
@@ -25,7 +25,10 @@ function echo_help() {
     echo "ASTRO_DEPLOYMENT_ID           Astro cloud Deployment id"
     echo "ASTRO_KEY_ID       Astro cloud service account API key id"
     echo "ASTRO_KEY_SECRET   Astro cloud service account API key secret"
-    echo "bash deploy.sh <ASTRO_DOCKER_REGISTRY> <ASTRO_ORGANIZATION_ID>  <ASTRO_DEPLOYMENT_ID> <ASTRO_KEY_ID> <ASTRO_KEY_SECRET>"
+    echo "ASTRO_DEPLOYMENT_ID_SINGLE_WORKER           Astro cloud Deployment id"
+    echo "ASTRO_KEY_ID_SINGLE_WORKER       Astro cloud service account for Single Worker API key id"
+    echo "ASTRO_KEY_SECRET_SINGLE_WORKER   Astro cloud service account for Single Worker API key secret"
+    echo "bash deploy.sh <ASTRO_DOCKER_REGISTRY> <ASTRO_ORGANIZATION_ID>  <ASTRO_DEPLOYMENT_ID> <ASTRO_KEY_ID> <ASTRO_KEY_SECRET> <ASTRO_DEPLOYMENT_ID_SINGLE_WORKER> <ASTRO_KEY_ID_SINGLE_WORKER> <ASTRO_KEY_SECRET_SINGLE_WORKER>"
 }
 
 if [ "$1" == "-h" ]; then
@@ -45,9 +48,68 @@ ASTRO_ORGANIZATION_ID=$2
 ASTRO_DEPLOYMENT_ID=$3
 ASTRO_KEY_ID=$4
 ASTRO_KEY_SECRET=$5
+ASTRO_DEPLOYMENT_ID_SINGLE_WORKER=$6
+ASTRO_KEY_ID_SINGLE_WORKER=$7
+ASTRO_KEY_SECRET_SINGLE_WORKER=$8
+MASTER_DAG_DOCKERFILE="Dockerfile"
+MASTER_DAG_MUTLI_WORKER_DOCKERFILE="Dockerfile.single_worker"
 
 clean
 
+function deploy(){
+    docker_registry_astro=$1
+    organization_id=$2
+    deployment_id=$3
+    key_id=$4
+    key_secret=$5
+    dockerfile=$6
+
+    # Build image and deploy
+    BUILD_NUMBER=$(awk 'BEGIN {srand(); print srand()}')
+    IMAGE_NAME=${docker_registry_astro}/${organization_id}/${deployment_id}:ci-${BUILD_NUMBER}
+    docker build --platform=linux/amd64 -t "${IMAGE_NAME}" -f "${SCRIPT_PATH}"/${dockerfile} "${SCRIPT_PATH}"
+    docker login "${docker_registry_astro}" -u "${key_id}" -p "${key_secret}"
+    docker push "${IMAGE_NAME}"
+
+    TOKEN=$( curl --location --request POST "https://auth.astronomer.io/oauth/token" \
+        --header "content-type: application/json" \
+        --data-raw "{
+            \"client_id\": \"$key_id\",
+            \"client_secret\": \"$key_secret\",
+            \"audience\": \"astronomer-ee\",
+            \"grant_type\": \"client_credentials\"}" | jq -r '.access_token' )
+
+    # Create the Image
+    echo "get image id"
+    IMAGE=$( curl --location --request POST "https://api.astronomer.io/hub/v1" \
+        --header "Authorization: Bearer $TOKEN" \
+        --header "Content-Type: application/json" \
+        --data-raw "{
+            \"query\" : \"mutation imageCreate(\n    \$input: ImageCreateInput!\n) {\n    imageCreate (\n    input: \$input\n) {\n    id\n    tag\n    repository\n    digest\n    env\n    labels\n    deploymentId\n  }\n}\",
+            \"variables\" : {
+                \"input\" : {
+                    \"deploymentId\" : \"$deployment_id\",
+                    \"tag\" : \"ci-$BUILD_NUMBER\"
+                    }
+                }
+            }" | jq -r '.data.imageCreate.id')
+    # Deploy the Image
+    echo "deploy image"
+    curl --location --request POST "https://api.astronomer.io/hub/v1" \
+            --header "Authorization: Bearer $TOKEN" \
+            --header "Content-Type: application/json" \
+            --data-raw "{
+                \"query\" : \"mutation imageDeploy(\n    \$input: ImageDeployInput!\n  ) {\n    imageDeploy(\n      input: \$input\n    ) {\n      id\n      deploymentId\n      digest\n      env\n      labels\n      name\n      tag\n      repository\n    }\n}\",
+                \"variables\" : {
+                    \"input\" : {
+                        \"id\" : \"$IMAGE\",
+                        \"tag\" : \"ci-$BUILD_NUMBER\",
+                        \"repository\" : \"images.astronomer.cloud/$organization_id/$deployment_id\"
+                        }
+                    }
+            }"
+
+}
 
 # Copy source files
 mkdir "${SCRIPT_PATH}"/python-sdk
@@ -59,49 +121,8 @@ cp -r "${PROJECT_PATH}"/example_dags "${SCRIPT_PATH}"/example_dags
 cp -r "${PROJECT_PATH}"/tests/data "${SCRIPT_PATH}"/tests/data
 
 
-# Build image and deploy
-BUILD_NUMBER=$(awk 'BEGIN {srand(); print srand()}')
-IMAGE_NAME=${ASTRO_DOCKER_REGISTRY}/${ASTRO_ORGANIZATION_ID}/${ASTRO_DEPLOYMENT_ID}:ci-${BUILD_NUMBER}
-docker build --platform=linux/amd64 -t "${IMAGE_NAME}" -f "${SCRIPT_PATH}"/Dockerfile "${SCRIPT_PATH}"
-docker login "${ASTRO_DOCKER_REGISTRY}" -u "${ASTRO_KEY_ID}" -p "${ASTRO_KEY_SECRET}"
-docker push "${IMAGE_NAME}"
+deploy $ASTRO_DOCKER_REGISTRY $ASTRO_ORGANIZATION_ID $ASTRO_DEPLOYMENT_ID $ASTRO_KEY_ID $ASTRO_KEY_SECRET $MASTER_DAG_DOCKERFILE
 
-TOKEN=$( curl --location --request POST "https://auth.astronomer.io/oauth/token" \
-      --header "content-type: application/json" \
-      --data-raw "{
-          \"client_id\": \"$ASTRO_KEY_ID\",
-          \"client_secret\": \"$ASTRO_KEY_SECRET\",
-          \"audience\": \"astronomer-ee\",
-          \"grant_type\": \"client_credentials\"}" | jq -r '.access_token' )
-
-# Step 5. Create the Image
-echo "get image id"
-IMAGE=$( curl --location --request POST "https://api.astronomer.io/hub/v1" \
-      --header "Authorization: Bearer $TOKEN" \
-      --header "Content-Type: application/json" \
-      --data-raw "{
-          \"query\" : \"mutation imageCreate(\n    \$input: ImageCreateInput!\n) {\n    imageCreate (\n    input: \$input\n) {\n    id\n    tag\n    repository\n    digest\n    env\n    labels\n    deploymentId\n  }\n}\",
-          \"variables\" : {
-              \"input\" : {
-                  \"deploymentId\" : \"$ASTRO_DEPLOYMENT_ID\",
-                  \"tag\" : \"ci-$BUILD_NUMBER\"
-                  }
-              }
-          }" | jq -r '.data.imageCreate.id')
-# Step 6. Deploy the Image
-echo "deploy image"
-curl --location --request POST "https://api.astronomer.io/hub/v1" \
-        --header "Authorization: Bearer $TOKEN" \
-        --header "Content-Type: application/json" \
-        --data-raw "{
-            \"query\" : \"mutation imageDeploy(\n    \$input: ImageDeployInput!\n  ) {\n    imageDeploy(\n      input: \$input\n    ) {\n      id\n      deploymentId\n      digest\n      env\n      labels\n      name\n      tag\n      repository\n    }\n}\",
-            \"variables\" : {
-                \"input\" : {
-                    \"id\" : \"$IMAGE\",
-                    \"tag\" : \"ci-$BUILD_NUMBER\",
-                    \"repository\" : \"images.astronomer.cloud/$ASTRO_ORGANIZATION_ID/$ASTRO_DEPLOYMENT_ID\"
-                    }
-                }
-          }"
+deploy $ASTRO_DOCKER_REGISTRY $ASTRO_ORGANIZATION_ID $ASTRO_DEPLOYMENT_ID_SINGLE_WORKER $ASTRO_KEY_ID_SINGLE_WORKER $ASTRO_KEY_SECRET_SINGLE_WORKER $MASTER_DAG_MUTLI_WORKER_DOCKERFILE
 
 clean